Refactored identity MongoDbPart to its own class for better recognition what is required
