+++ 9_MongoRepository/src/QuickstartIdentityServer/MongoDbStartup.cs
@@ -1,0 +1,94 @@
+﻿using System;
+using IdentityServer4.Models;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+using MongoDB.Bson.Serialization;
+using QuickstartIdentityServer.Quickstart.Interface;
+
+namespace QuickstartIdentityServer
+{
+    public static class MongoDbStartup
+    {
+        /// <summary>
+        /// Adds the support for MongoDb Persistance for all identityserver stored
+        /// </summary>
+        /// <remarks><![CDATA[
+        /// It implements and used mongodb collections for:
+        /// - Clients
+        /// - IdentityResources
+        /// - ApiResource
+        /// ]]></remarks>
+        public static void UseMongoDbForIdentityServer(this IApplicationBuilder app)
+        {
+            var createdNewRepository = false;
+            var repository = app.ApplicationServices.GetService<IRepository>();
+
+            //  --Client
+            if (!repository.CollectionExists<Client>())
+            {
+                foreach (var client in Config.GetClients())
+                {
+                    repository.Add(client);
+                }
+                createdNewRepository = true;
+            }
+
+            //  --IdentityResource
+            if (!repository.CollectionExists<IdentityResource>())
+            {
+                foreach (var res in Config.GetIdentityResources())
+                {
+                    repository.Add(res);
+                }
+                createdNewRepository = true;
+            }
+
+
+            //  --ApiResource
+            if (!repository.CollectionExists<ApiResource>())
+            {
+                foreach (var api in Config.GetApiResources())
+                {
+                    repository.Add(api);
+                }
+                createdNewRepository = true;
+            }
+
+            // If it's a new Repository (database), need to restart the website to configure Mongo to ignore Extra Elements.
+            if (createdNewRepository)
+            {
+                var newRepositoryMsg =
+                    $"Mongo Repository created/populated! Please restart you website, so Mongo driver will be configured  to ignore Extra Elements.";
+                throw new Exception(newRepositoryMsg);
+            }
+        }
+
+        /// <summary>
+        /// Configure Classes to ignore Extra Elements (e.g. _Id) when deserializing
+        /// As we are using "IdentityServer4.Models" we cannot add something like "[BsonIgnore]"
+        /// </summary>
+        private static void ConfigureMongoDriver2IgnoreExtraElements()
+        {
+            BsonClassMap.RegisterClassMap<Client>(cm =>
+            {
+                cm.AutoMap();
+                cm.SetIgnoreExtraElements(true);
+            });
+            BsonClassMap.RegisterClassMap<IdentityResource>(cm =>
+            {
+                cm.AutoMap();
+                cm.SetIgnoreExtraElements(true);
+            });
+            BsonClassMap.RegisterClassMap<ApiResource>(cm =>
+            {
+                cm.AutoMap();
+                cm.SetIgnoreExtraElements(true);
+            });
+            BsonClassMap.RegisterClassMap<PersistedGrant>(cm =>
+            {
+                cm.AutoMap();
+                cm.SetIgnoreExtraElements(true);
+            });
+        }
+    }
+}
