+++ 9_MongoRepository/src/QuickstartIdentityServer/Startup.cs
@@ -93,80 +93,7 @@ namespace QuickstartIdentityServer
         }
 
 
-        private static void InitializeDatabase(IApplicationBuilder app)
-        {
-            bool createdNewRepository = false;
-            var repository = app.ApplicationServices.GetService<IRepository>();
-
-            //  --Client
-            if (!repository.CollectionExists<Client>())
-            {
-                foreach (var client in Config.GetClients())
-                {
-                    repository.Add<Client>(client);
-                }
-                createdNewRepository = true;
-            }
-
-            //  --IdentityResource
-            if (!repository.CollectionExists<IdentityResource>())
-            {
-                foreach (var res in Config.GetIdentityResources())
-                {
-                    repository.Add<IdentityResource>(res);
-                }
-                createdNewRepository = true;
-            }
-
-
-            //  --ApiResource
-            if (!repository.CollectionExists<ApiResource>())
-            {
-                foreach (var api in Config.GetApiResources())
-                {
-                    repository.Add<ApiResource>(api);
-                }
-                createdNewRepository = true;
-            }
-
-            // If it's a new Repository (database), need to restart the website to configure Mongo to ignore Extra Elements.
-            if (createdNewRepository)
-            {
-                var newRepositoryMsg = $"Mongo Repository created/populated! Please restart you website, so Mongo driver will be configured  to ignore Extra Elements.";
-                throw new Exception(newRepositoryMsg);
-            }
-        }
-
-        /// <summary>
-        /// Configure Classes to ignore Extra Elements (e.g. _Id) when deserializing
-        /// As we are using "IdentityServer4.Models" we cannot add something like "[BsonIgnore]"
-        /// </summary>
-        private static void ConfigureMongoDriver2IgnoreExtraElements()
-        {
-            BsonClassMap.RegisterClassMap<Client>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetIgnoreExtraElements(true);
-            });
-            BsonClassMap.RegisterClassMap<IdentityResource>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetIgnoreExtraElements(true);
-            });
-            BsonClassMap.RegisterClassMap<ApiResource>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetIgnoreExtraElements(true);
-            });
-            BsonClassMap.RegisterClassMap<PersistedGrant>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetIgnoreExtraElements(true);
-            });
-
-
-
-        }
+     
 
     }
 }
